+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Enh #163: Add `workers` option to `serve` command with default of two workers under Linux (@xepozz)
 - Enh #162: Explicitly add transitive dependencies `psr/event-dispatcher` and `psr/log` (@vjik)
+- Bug #168: Fix executing the `list` command with namespace (@vjik)
 
 ## 1.3.0 July 29, 2022